+++ config/script_test.go
@@ -4,164 +4,167 @@
 package config_test
 
 import (
+	"fmt"
 	"strings"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	hhcl "github.com/hashicorp/hcl/v2"
 	"github.com/madlambda/spells/assert"
 	"github.com/terramate-io/terramate/config"
 	"github.com/terramate-io/terramate/errors"
 	"github.com/terramate-io/terramate/hcl"
-	"github.com/terramate-io/terramate/hcl/ast"
 	"github.com/terramate-io/terramate/hcl/eval"
 	"github.com/terramate-io/terramate/hcl/info"
+	"github.com/terramate-io/terramate/project"
 	"github.com/terramate-io/terramate/stdlib"
 	"github.com/terramate-io/terramate/test"
+	errtest "github.com/terramate-io/terramate/test/errors"
 
-	. "github.com/terramate-io/terramate/test/hclutils"
-	. "github.com/terramate-io/terramate/test/hclutils/info"
+	. "github.com/terramate-io/terramate/test/hclwrite/hclutils"
 	"github.com/zclconf/go-cty/cty"
 )
+
+type scriptTestcase struct {
+	name    string
+	config  fmt.Stringer
+	globals map[string]cty.Value
+	want    config.Script
+	wantErr error
+}
 
 func TestScriptEval(t *testing.T) {
 	t.Parallel()
 
-	makeAttribute := func(t *testing.T, name string, expr string) *ast.Attribute {
-		t.Helper()
-		return &ast.Attribute{
-			Attribute: &hhcl.Attribute{
-				Name: name,
-				Expr: test.NewExpr(t, expr),
-			},
-		}
-	}
-
-	makeCommand := func(t *testing.T, expr string) *hcl.Command {
-		t.Helper()
-		attr := makeAttribute(t, "command", expr)
-		parsed := hcl.Command(*attr)
-		return &parsed
-	}
-
-	makeCommands := func(t *testing.T, expr string) *hcl.Commands {
-		t.Helper()
-		attr := makeAttribute(t, "commands", expr)
-		parsed := hcl.Commands(*attr)
-		return &parsed
-	}
-
 	labels := []string{"some", "label"}
 
-	type testcase struct {
-		name    string
-		script  hcl.Script
-		globals map[string]cty.Value
-		want    config.Script
-		wantErr error
-	}
-
-	tcases := []testcase{
+	tcases := []scriptTestcase{
 		{
 			name: "no description attribute",
-			script: hcl.Script{
-				Labels: labels,
-			},
+			config: Script(
+				Labels(labels...),
+				Block("job",
+					Command("echo", "hello"),
+				),
+			),
 			want: config.Script{
 				Labels: labels,
+				Jobs: []config.ScriptJob{
+					{Cmd: &config.ScriptCmd{Args: []string{"echo", "hello"}}},
+				},
 			},
 		},
 		{
 			name: "description attribute wrong type",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `666`),
-			},
+			config: Script(
+				Labels(labels...),
+				Number("description", 666),
+				Block("job", Command("echo", "hello")),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidType),
 		},
 		{
 			name: "description attribute with functions",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `tm_upper("some description")`),
-			},
+			config: Script(
+				Labels(labels...),
+				Expr("description", `tm_upper("some description")`),
+				Block("job", Command("echo", "hello")),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Description: "SOME DESCRIPTION",
+				Jobs: []config.ScriptJob{
+					{Cmd: &config.ScriptCmd{Args: []string{"echo", "hello"}}},
+				},
 			},
 		},
 		{
 			name: "description attribute with functions and globals",
-			script: hcl.Script{
-				Range:       Range("script.tm", Start(1, 1, 0), End(3, 2, 37)),
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `tm_upper(global.some_string_var)`),
-			},
+			config: Script(
+				Labels(labels...),
+				Expr("description", `tm_upper(global.some_string_var)`),
+				Block("job", Command("echo", "hello")),
+			),
 			globals: map[string]cty.Value{
 				"some_string_var": cty.StringVal("terramate"),
 			},
 			want: config.Script{
 				Labels:      labels,
 				Description: "TERRAMATE",
+				Jobs: []config.ScriptJob{
+					{Cmd: &config.ScriptCmd{Args: []string{"echo", "hello"}}},
+				},
 			},
 		},
 		{
 			name: "name attribute with wrong type",
-			script: hcl.Script{
-				Labels:      labels,
-				Name:        makeAttribute(t, "name", `666`),
-				Description: makeAttribute(t, "description", `"some desc"`),
-			},
+			config: Script(
+				Labels(labels...),
+				Number("name", 666),
+				Str("description", "some desc"),
+				Block("job", Command("echo", "hello")),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidType),
 		},
 		{
 			name: "name attribute with string",
-			script: hcl.Script{
-				Labels:      labels,
-				Name:        makeAttribute(t, "name", `"some name"`),
-				Description: makeAttribute(t, "description", `"some desc"`),
-			},
+			config: Script(
+				Labels(labels...),
+				Str("name", "some name"),
+				Str("description", "some desc"),
+				Block("job", Command("echo", "hello")),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Name:        "some name",
 				Description: "some desc",
+				Jobs: []config.ScriptJob{
+					{Cmd: &config.ScriptCmd{Args: []string{"echo", "hello"}}},
+				},
 			},
 		},
 		{
 			name: "name attribute exceeds maximum allowed characters - truncation",
-			script: hcl.Script{
-				Labels:      labels,
-				Name:        makeAttribute(t, "name", `"`+strings.Repeat("A", 150)+`"`),
-				Description: makeAttribute(t, "description", `"some desc"`),
-			},
+			config: Script(
+				Labels(labels...),
+				Str("name", strings.Repeat("A", 150)),
+				Str("description", "some desc"),
+				Block("job", Command("echo", "hello")),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Name:        strings.Repeat("A", 128),
 				Description: "some desc",
+				Jobs: []config.ScriptJob{
+					{Cmd: &config.ScriptCmd{Args: []string{"echo", "hello"}}},
+				},
 			},
 		},
 		{
 			name: "name attribute with functions",
-			script: hcl.Script{
-				Labels:      labels,
-				Name:        makeAttribute(t, "name", `tm_upper("some name")`),
-				Description: makeAttribute(t, "description", `"some desc"`),
-			},
+			config: Script(
+				Labels(labels...),
+				Expr("name", `tm_upper("some name")`),
+				Str("description", "some desc"),
+				Block("job", Command("echo", "hello")),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Name:        "SOME NAME",
 				Description: "some desc",
+				Jobs: []config.ScriptJob{
+					{Cmd: &config.ScriptCmd{Args: []string{"echo", "hello"}}},
+				},
 			},
 		},
 		{
 			name: "name attribute with interpolation, functions and globals",
-			script: hcl.Script{
-				Range:       Range("script.tm", Start(1, 1, 0), End(3, 2, 37)),
-				Labels:      labels,
-				Name:        makeAttribute(t, "name", `"my name is ${tm_upper(global.name_var)}!!!"`),
-				Description: makeAttribute(t, "description", `"some desc"`),
-			},
+			config: Script(
+				Labels(labels...),
+				Expr("name", `"my name is ${tm_upper(global.name_var)}!!!"`),
+				Str("description", "some desc"),
+				Block("job", Command("echo", "hello")),
+			),
 			globals: map[string]cty.Value{
 				"name_var": cty.StringVal("terramate"),
 			},
@@ -169,19 +172,20 @@ func TestScriptEval(t *testing.T) {
 				Labels:      labels,
 				Name:        "my name is TERRAMATE!!!",
 				Description: "some desc",
+				Jobs: []config.ScriptJob{
+					{Cmd: &config.ScriptCmd{Args: []string{"echo", "hello"}}},
+				},
 			},
 		},
 		{
-			name: "command attribute wrong type",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Command: makeCommand(t, `"echo"`),
-					},
-				},
-			},
+			name: "command attribute wrong type - must be a list",
+			config: Script(
+				Labels(labels...),
+				Str("description", "some desc"),
+				Block("job",
+					Str("command", "echo"),
+				),
+			),
 			globals: map[string]cty.Value{
 				"some_string_var": cty.StringVal("terramate"),
 			},
@@ -189,28 +193,24 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "command attribute wrong element type",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Command: makeCommand(t, `["echo", 666]`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some desc"),
+				Block("job",
+					Expr("command", `["echo", 666]`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidTypeCommand),
 		},
 		{
 			name: "command attribute with functions and globals",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Command: makeCommand(t, `["echo", tm_upper("hello ${global.some_string_var}")]`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("command", `["echo", tm_upper("hello ${global.some_string_var}")]`),
+				),
+			),
 			globals: map[string]cty.Value{
 				"some_string_var": cty.StringVal("terramate"),
 			},
@@ -228,15 +228,12 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "command attribute with list type",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Command: makeCommand(t, `true ? tm_concat(["echo"], ["something"]) : []`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("command", `true ? tm_concat(["echo"], ["something"]) : []`)),
+			),
 			globals: map[string]cty.Value{
 				"some_string_var": cty.StringVal("terramate"),
 			},
@@ -254,15 +251,13 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "commands attribute with list type",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `true ? tm_concat([["echo", "something"]], [["echo", "other", "thing"]]) : []`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `true ? tm_concat([["echo", "something"]], [["echo", "other", "thing"]]) : []`),
+				),
+			),
 			globals: map[string]cty.Value{
 				"some_string_var": cty.StringVal("terramate"),
 			},
@@ -285,15 +280,13 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "command with first item interpolated",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Command: makeCommand(t, `["${global.some_command_name}", "--version"]`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("command", `["${global.some_command_name}", "--version"]`),
+				),
+			),
 			globals: map[string]cty.Value{
 				"some_command_name": cty.StringVal("ls"),
 			},
@@ -311,15 +304,13 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "commands attribute wrong type",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `"echo"`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Str("commands", "echo"),
+				),
+			),
 			globals: map[string]cty.Value{
 				"some_string_var": cty.StringVal("terramate"),
 			},
@@ -327,81 +318,65 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "commands attribute wrong element type",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello"],
-							666,
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Str("commands", "test"),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidTypeCommands),
 		},
 		{
 			name: "commands evaluating to empty list",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `[]`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[]`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptEmptyCmds),
 		},
 		{
 			name: "commands item evaluating to empty list",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `[
-							["echo", "hello"],
-							[],
-							["echo", "other"]
-						]`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello"],
+								[],
+								["echo", "other"]
+							]`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptEmptyCmds),
 		},
 		{
 			name: "command evaluating to empty list",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Command: makeCommand(t, `[]`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("command", `[]`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptEmptyCmds),
 		},
 		{
 			name: "commands attribute with functions and globals",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", tm_upper("hello ${global.some_string_var}")],
-							["stat", "."],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", tm_upper("hello ${global.some_string_var}")],
+								["stat", "."],
+							  ]
+							`),
+				),
+			),
 			globals: map[string]cty.Value{
 				"some_string_var": cty.StringVal("terramate"),
 			},
@@ -420,29 +395,24 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "multiple jobs",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Name: makeAttribute(t, "name", `"job name"`),
-						Commands: makeCommands(t, `
-						  [
-							["echo", tm_upper("hello ${global.some_string_var}")],
-							["stat", "."],
-						  ]
-						`),
-					},
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", tm_upper("hello ${global.some_string_var}")],
-							["ls", "-l"],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Str("name", "job name"),
+					Expr("commands", `[
+								["echo", tm_upper("hello ${global.some_string_var}")],
+								["stat", "."],
+							  ]`),
+				),
+				Block("job",
+					Expr("commands", `[
+								["echo", tm_upper("hello ${global.some_string_var}")],
+								["ls", "-l"],
+							  ]
+							`),
+				),
+			),
 			globals: map[string]cty.Value{
 				"some_string_var": cty.StringVal("terramate"),
 			},
@@ -468,20 +438,14 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "job.name attribute exceeds maximum allowed characters - truncation",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some desc"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Name: makeAttribute(t, "name", `"`+strings.Repeat("A", 150)+`"`),
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello"],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some desc"),
+				Block("job",
+					Str("name", strings.Repeat("A", 150)),
+					Expr("commands", `[["echo", "hello"]]`),
+				),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Description: "some desc",
@@ -497,20 +461,14 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "job.description attribute exceeds maximum allowed characters - truncation",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some desc"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Description: makeAttribute(t, "description", `"`+strings.Repeat("A", config.MaxScriptDescRunes+100)+`"`),
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello"],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some desc"),
+				Block("job",
+					Str("description", strings.Repeat("A", config.MaxScriptDescRunes+100)),
+					Expr("commands", `[["echo", "hello"]]`),
+				),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Description: "some desc",
@@ -526,65 +484,56 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "command options with sync_deployment and sync_preview",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello", {
-								sync_deployment = true
-								sync_preview = true
-								terraform_plan_file = "plan_a"
-							}],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello", {
+									sync_deployment = true
+									sync_preview = true
+									terraform_plan_file = "plan_a"
+								}],
+							  ]
+							`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidCmdOptions),
 		},
 		{
 			name: "command options with invalid layer",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello", {
-								sync_preview = true
-								terraform_plan_file = "plan_a"
-								layer = "a+b"
-							}],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello", {
+									sync_preview = true
+									terraform_plan_file = "plan_a"
+									layer = "a+b"
+								}],
+							  ]
+							`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidCmdOptions),
 		},
 		{
 			name: "command options with sync_preview + planfile + layer",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello", {
-								sync_preview = true
-								terraform_plan_file = "plan_a"
-								layer = "staging"
-							}],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello", {
+									sync_preview = true
+									terraform_plan_file = "plan_a"
+									layer = "staging"
+								}],
+							  ]
+							`),
+				),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Description: "some description",
@@ -607,30 +556,25 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "command options",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello", {
-								sync_deployment = false
-								terraform_plan_file = "plan_a"
-							}],
-						  ]
-						`),
-					},
-					{
-						Command: makeCommand(t, `
-							["echo", "hello", {
-								sync_deployment = true
-								terraform_plan_file = "plan_b"
-							}]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello", {
+									sync_deployment = false
+									terraform_plan_file = "plan_a"
+								}],
+							  ]
+							`),
+				),
+				Block("job",
+					Expr("command", `["echo", "hello", {
+									sync_deployment = true
+									terraform_plan_file = "plan_b"
+								}]`),
+				),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Description: "some description",
@@ -660,23 +604,20 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "command options with planfile + terragrunt",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello", {
-								sync_deployment = true
-								terraform_plan_file = "plan_a"
-								terragrunt = true
-							}],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello", {
+									sync_deployment = true
+									terraform_plan_file = "plan_a"
+									terragrunt = true
+								}],
+							  ]
+							`),
+				),
+			),
 			want: config.Script{
 				Labels:      labels,
 				Description: "some description",
@@ -698,85 +639,70 @@ func TestScriptEval(t *testing.T) {
 		},
 		{
 			name: "invalid command option",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello", {
-								sync_deploymenttttttt = false
-							}],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello", {
+									sync_deploymenttttttt = false
+								}],
+							  ]
+							`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidCmdOptions),
 		},
 		{
 			name: "invalid command options object",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello", ["list"]],
-						  ]
-						`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello", ["list"]],
+							  ]
+							`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidTypeCommand),
 		},
 		{
-			name: "invalid command option type",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-						  [
-							["echo", "hello", {
-								sync_deployment = "false"
-							}],
-						  ]
-						`),
-					},
-				},
-			},
+			name: "invalid command option type - option must be boolean",
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+								["echo", "hello", {
+									sync_deployment = "false"
+								}],
+							  ]
+							`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidCmdOptions),
 		},
 		{
 			name: "multiple sync_deployments",
-			script: hcl.Script{
-				Labels:      labels,
-				Description: makeAttribute(t, "description", `"some description"`),
-				Jobs: []*hcl.ScriptJob{
-					{
-						Commands: makeCommands(t, `
-							  [
-								["echo", "hello", {
-									sync_deployment = true
-									terraform_plan_file = "plan_a"
-								}],
-							  ]
-							`),
-					},
-					{
-						Command: makeCommand(t, `
-								["echo", "hello", {
-									sync_deployment = true
-									terraform_plan_file = "plan_a"
-								}]
-							`),
-					},
-				},
-			},
+			config: Script(
+				Labels(labels...),
+				Str("description", "some description"),
+				Block("job",
+					Expr("commands", `[
+									["echo", "hello", {
+										sync_deployment = true
+										terraform_plan_file = "plan_a"
+									}],
+								  ]`),
+				),
+				Block("job",
+					Expr("commands", `[["echo", "hello", {
+										sync_deployment = true
+										terraform_plan_file = "plan_a"
+									}]]`),
+				),
+			),
 			wantErr: errors.E(config.ErrScriptInvalidCmdOptions),
 		},
 	}
@@ -785,15 +711,50 @@ func TestScriptEval(t *testing.T) {
 		tcase := tcase
 		t.Run(tcase.name, func(t *testing.T) {
 			t.Parallel()
-			hclctx := eval.NewContext(stdlib.Functions(test.TempDir(t)))
-			hclctx.SetNamespace("global", tcase.globals)
-
-			got, err := config.EvalScript(hclctx, tcase.script)
-			assert.IsError(t, err, tcase.wantErr)
-			// ignoring info.Range comparisons for now
-			if diff := cmp.Diff(tcase.want, got, cmpopts.IgnoreUnexported(info.Range{})); diff != "" {
-				t.Fatalf("unexpected result\n%s", diff)
-			}
+			testScriptEval(t, tcase)
 		})
+	}
+}
+
+func testScriptEval(t *testing.T, tcase scriptTestcase) {
+	t.Helper()
+	tempdir := test.TempDir(t)
+	test.AppendFile(t, tempdir, "stack.tm", Block("stack").String())
+	test.AppendFile(t, tempdir, "script.tm", tcase.config.String())
+	test.AppendFile(t, tempdir, "terramate.tm", Terramate(
+		Config(
+			Expr("experiments", `["scripts"]`),
+		),
+	).String())
+
+	cfg, err := config.LoadRoot(tempdir)
+	if errors.IsAnyKind(tcase.wantErr, hcl.ErrHCLSyntax, hcl.ErrTerramateSchema) {
+		errtest.Assert(t, err, tcase.wantErr)
+		return
+	}
+
+	assert.NoError(t, err)
+
+	rootTree, ok := cfg.Lookup(project.NewPath("/"))
+	if !ok {
+		panic("root tree not found")
+	}
+
+	st, err := rootTree.Stack()
+	assert.NoError(t, err)
+	hclctx := eval.NewContext(stdlib.Functions(tempdir))
+	hclctx.SetNamespace("global", tcase.globals)
+	runtime := cfg.Runtime()
+	runtime.Merge(st.RuntimeValues(cfg))
+	hclctx.SetNamespace("terramate", runtime)
+
+	if len(rootTree.Node.Scripts) != 1 {
+		panic("test expects one script")
+	}
+	got, err := config.EvalScript(hclctx, *rootTree.Node.Scripts[0])
+	assert.IsError(t, err, tcase.wantErr)
+	// ignoring info.Range comparisons for now
+	if diff := cmp.Diff(tcase.want, got, cmpopts.IgnoreUnexported(info.Range{})); diff != "" {
+		t.Fatalf("unexpected result\n%s", diff)
 	}
 }