+++ hcl/hcl_test.go
@@ -1667,7 +1667,9 @@ func TestHCLParserTerramateBlocksMerging(t *testing.T) {
 }
 
 func testParser(t *testing.T, tc testcase) {
+	t.Helper()
 	t.Run(tc.name, func(t *testing.T) {
+		t.Helper()
 		t.Parallel()
 		configsDir := test.TempDir(t)
 		for _, inputConfigFile := range tc.input {